:sparkles: (feature/takepic): Loading prompt during scan
